+++ src/frontend/apps/e2e/__tests__/app-impress/doc-search.spec.ts
@@ -1,6 +1,7 @@
 import { expect, test } from '@playwright/test';
 
 import { createDoc, randomName, verifyDocName } from './utils-common';
+import { createRootSubPage } from './utils-sub-pages';
 
 test.beforeEach(async ({ page }) => {
   await page.goto('/');
@@ -98,39 +99,46 @@ test.describe('Document search', () => {
     ).toBeHidden();
   });
 
-  test("it checks we don't see filters in search modal", async ({ page }) => {
-    const searchButton = page
-      .getByTestId('left-panel-desktop')
-      .getByRole('button', { name: 'search' });
-
-    await expect(searchButton).toBeVisible();
-    await page.getByRole('button', { name: 'search', exact: true }).click();
-    await expect(
-      page.getByRole('combobox', { name: 'Quick search input' }),
-    ).toBeVisible();
-    await expect(page.getByTestId('doc-search-filters')).toBeHidden();
-  });
-});
-
-test.describe('Sub page search', () => {
   test('it check the presence of filters in search modal', async ({
     page,
     browserName,
   }) => {
-    await page.goto('/');
-    const [doc1Title] = await createDoc(
-      page,
-      'My sub page search',
-      browserName,
-      1,
-    );
-    await verifyDocName(page, doc1Title);
+    // Doc grid filters are not visible
     const searchButton = page
       .getByTestId('left-panel-desktop')
-      .getByRole('button', { name: 'search' });
-    await searchButton.click();
+      .getByRole('button', { name: 'search', exact: true });
+
     const filters = page.getByTestId('doc-search-filters');
+
+    await searchButton.click();
+    await expect(
+      page.getByRole('combobox', { name: 'Quick search input' }),
+    ).toBeVisible();
+    await expect(filters).toBeHidden();
+
+    await page.getByRole('button', { name: 'close' }).click();
+
+    // Create a doc without children for the moment
+    // and check that filters are not visible
+    const [doc1Title] = await createDoc(page, 'My page search', browserName, 1);
+    await verifyDocName(page, doc1Title);
+
+    await searchButton.click();
+    await expect(
+      page.getByRole('combobox', { name: 'Quick search input' }),
+    ).toBeVisible();
+    await expect(filters).toBeHidden();
+
+    await page.getByRole('button', { name: 'close' }).click();
+
+    // Create a sub page
+    // and check that filters are visible
+    await createRootSubPage(page, browserName, 'My sub page search');
+
+    await searchButton.click();
+
     await expect(filters).toBeVisible();
+
     await filters.click();
     await filters.getByRole('button', { name: 'Current doc' }).click();
     await expect(
@@ -139,43 +147,70 @@ test.describe('Sub page search', () => {
     await expect(
       page.getByRole('menuitem', { name: 'Current doc' }),
     ).toBeVisible();
-    await page.getByRole('menuitem', { name: 'Current doc' }).click();
+    await page.getByRole('menuitem', { name: 'All docs' }).click();
 
     await expect(page.getByRole('button', { name: 'Reset' })).toBeVisible();
   });
 
   test('it searches sub pages', async ({ page, browserName }) => {
-    await page.goto('/');
-
-    const [doc1Title] = await createDoc(
+    // First doc
+    const [firstDocTitle] = await createDoc(
       page,
-      'My sub page search',
+      'My first sub page search',
       browserName,
       1,
     );
-    await verifyDocName(page, doc1Title);
+    await verifyDocName(page, firstDocTitle);
+
+    // Create a new doc - for the moment without children
     await page.getByRole('button', { name: 'New doc' }).click();
     await verifyDocName(page, '');
     await page.getByRole('textbox', { name: 'doc title input' }).click();
     await page
       .getByRole('textbox', { name: 'doc title input' })
       .press('ControlOrMeta+a');
-    const [randomDocName] = randomName('doc-sub-page', browserName, 1);
+    const [secondDocTitle] = randomName(
+      'My second sub page search',
+      browserName,
+      1,
+    );
     await page
       .getByRole('textbox', { name: 'doc title input' })
-      .fill(randomDocName);
+      .fill(secondDocTitle);
+
     const searchButton = page
       .getByTestId('left-panel-desktop')
       .getByRole('button', { name: 'search' });
 
     await searchButton.click();
-    await expect(
-      page.getByRole('button', { name: 'Current doc' }),
-    ).toBeVisible();
     await page.getByRole('combobox', { name: 'Quick search input' }).click();
     await page
       .getByRole('combobox', { name: 'Quick search input' })
-      .fill('sub');
-    await expect(page.getByLabel(randomDocName)).toBeVisible();
+      .fill('sub page search');
+
+    // Expect to find the first doc
+    await expect(
+      page.getByRole('presentation').getByLabel(firstDocTitle),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('presentation').getByLabel(secondDocTitle),
+    ).toBeVisible();
+
+    await page.getByRole('button', { name: 'close' }).click();
+
+    // Create a sub page
+    await createRootSubPage(page, browserName, secondDocTitle);
+    await searchButton.click();
+    await page
+      .getByRole('combobox', { name: 'Quick search input' })
+      .fill('sub page search');
+
+    // Now there is a sub page - expect to have the focus on the current doc
+    await expect(
+      page.getByRole('presentation').getByLabel(secondDocTitle),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('presentation').getByLabel(firstDocTitle),
+    ).toBeHidden();
   });
 });