+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useCollaboration';
 export * from './useCopyDocLink';
+export * from './useDocUtils';
 export * from './useIsCollaborativeEditable';
 export * from './useTrans';