+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/useDocUtils.tsx
@@ -1,9 +1,10 @@
 import { Doc } from '@/docs/doc-management';
 
-export const useTreeUtils = (doc: Doc) => {
+export const useDocUtils = (doc: Doc) => {
   return {
     isTopRoot: doc.depth === 1,
     isChild: doc.depth > 1,
+    hasChildren: doc.numchild > 0,
     isDesynchronized: !!(
       doc.ancestors_link_reach &&
       (doc.computed_link_reach !== doc.ancestors_link_reach ||